Refactored ComplexBoundingBox to CompoundBoundingBox and extracted into its own file.
